# `Array` performance improvements to reduce copying/copy_on_write calls