- Avoid temporary copy of `p_array` in `Array::append_array` when types match
- Call `ptrw()` once before looping in methods that return new `Arrays`, to avoid `copy_on_write` call for each item (`recursive_duplicate`, `slice`, `filter`, `map`)

Compared with gdscript below.  Most methods changed are usually around 10-25% faster, will vary depending on types and sizes of arrays.

```gdscript
func _ready() -> void:
	time("test_append_array_untyped")
	time("test_append_array_int_float")
	time("test_duplicate_deep")
	time("test_slice_int")
	time("test_slice_str")
	time("test_filter")
	time("test_map")
	time("test_map_lambda")

func test_append_array_untyped():
	var a := []
	var r := range(100)
	for i in 100000:
		a.append_array(r)

func test_append_array_int_float():
	var a : Array[float] = []
	var r := range(100)
	for i in 100000:
		a.append_array(r)

func test_duplicate_deep():
	var a := range(100).map(func(i): return range(i))
	for i in 10000:
		a.duplicate(true)
	
func test_slice_int():
	var a := range(100)
	for i in 1000000:
		a.slice(0, 20)
		
func test_slice_str():
	var a := range(100).map(str)
	for i in 1000000:
		a.slice(0, 20)

func test_filter():
	var a := range(1000)
	for i in 1000:
		a.filter(func(n): return n % 3 == 0)

func test_map():
	var a := range(1000)
	for i in 10000:
		a.map(sqrt)

func test_map_lambda():
	var a := range(1000)
	for i in 10000:
		a.map(func(n): return n * n)

func time(test_name : String):
	var start := Time.get_ticks_msec()
	call(test_name)
	var end := Time.get_ticks_msec()
	print("%s: %dms" % [test_name, end - start])
```

old
```
test_append_array_untyped: 251ms
test_append_array_int_float: 307ms
test_duplicate_deep: 1002ms
test_slice_int: 572ms
test_slice_str: 1150ms
test_filter: 150ms
test_map: 421ms
test_map_lambda: 1092ms
```

new
```
test_append_array_untyped: 186ms
test_append_array_int_float: 300ms
test_duplicate_deep: 793ms
test_slice_int: 455ms
test_slice_str: 1096ms
test_filter: 137ms
test_map: 333ms
test_map_lambda: 979ms
```

<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://docs.godotengine.org/en/latest/contributing/development/code_style_guidelines.html
-->